Normalize neural network input

+++ app/src/main/java/com/ps/simplepapersoccer/AI/neuralnetworkAI/INeuralNetworkController.kt
@@ -1,7 +1,7 @@
 package com.ps.simplepapersoccer.ai.neuralnetworkAI
 
 interface INeuralNetworkController<T> {
-    val inputs: List<Int>
+    val inputs: List<Float>
     val outputs: Int
     fun updateInputs()
     fun fitnessEvaluation(): Float
