+++ app/src/main/java/com/ps/simplepapersoccer/GameObjects/Game/Geometry/Abstraction/BaseNode.kt
@@ -3,7 +3,7 @@ package com.ps.simplepapersoccer.gameObjects.game.geometry.abstraction
 import com.ps.simplepapersoccer.gameObjects.game.geometry.TwoDimensionalPoint
 
 abstract class BaseNode(val coords: TwoDimensionalPoint) {
-    abstract fun normalizedIdentifierHashCode(): Int
+    abstract fun normalizedIdentifierHashCode(): Float
 
     abstract val getVisibleCoords: TwoDimensionalPoint
 }
